fix(ios): use different gesture tags
when there is more than one pie chart on the same page on iOS and the latest added pie chart slices get tapped the first pie chart's tap gesture handler gets triggered. this commit aims to fix this issue. (bar/line charts already have the logic that is added with this commit for pie and radar charts)